telegram_bot/requirements.txt
aiogram>=3.3.0
httpx==0.24.1
pydantic>=2.1.1
email-validator==2.3.0
python-dotenv==1.0.0
python-jose==3.3.0
cryptography==41.0.7
pytest==9.0.2
pytest-asyncio==1.3.0
